Change Github apis
The previous version of the apis expose two main functions:
`findLatestRelease` and `findLatestAndPrereleaseReleases`. The two functions,
however, are not what we meant to use.

Switch to two new main apis, `findLatestRelease` and `findLatestStableRelease`.
Both of the new functions return only one release based on if we want to
take pre-releases into account when finding a release or not.

In addition:
- Remove some redundant functions
- Edit name of some of the functions
- Update jsdoc of some of the functions
- Switch `hasAssetForChainType` implementation to a curried one, so that
  it can be used as a predicate more easily