Add permit health check parameter
RWT count should be more than the required number used to create a
commitment, if the event creation speed is faster than the reward period
you may need to lock more RSN and get more RWT to work properly. This
health check parameter checks the remaining RWT and notify the watcher
if it is less than the recommended amount.